Add book.html by upload
This is the book now page. The content of the page at the moment is more like a priceless page. There will be additional piece of code at the bottom with a book now form, that will allow to choose date and time for the appointment. The page contains the markup from index.html for the header and additional 3 tables with all the prices and services.