add book.js by upload
This is a javascript file for the Book.html page. It contains info about the services, but is added to the Dom via javascript. I used JS Class and then I used the class to create an object for each service. This has been later added to the Dom using a template literal.